+++ pkg/api/version/types/types.go
@@ -9,7 +9,7 @@ import (
 type AppVersion struct {
 	KOTSKinds  *kotsutil.KotsKinds `json:"kotsKinds"`
 	AppID      string              `json:"appId"`
-	Sequence   int64               `json:"sequence"`
+	Sequence   float64             `json:"sequence"`
 	Status     string              `json:"status"`
 	CreatedOn  time.Time           `json:"createdOn"`
 	DeployedAt *time.Time          `json:"deployedAt"`
