+++ pkg/print/versions.go
@@ -8,7 +8,7 @@ import (
 
 type AppVersionResponse struct {
 	VersionLabel string     `json:"versionLabel"`
-	Sequence     int64      `json:"sequence"`
+	Sequence     float64    `json:"sequence"`
 	CreatedOn    time.Time  `json:"createdOn"`
 	Status       string     `json:"status"`
 	DeployedAt   *time.Time `json:"deployedAt"`
